requirements.txt
requests
pandas
jsonschema
google-cloud-storage
google-cloud-bigquery
pandas-gbq
